Remove MSC version check for stdatomic and threads
While the check is technically correct (support for the c11 standard was
added in MSVC 1928), in order to use it the solution must set the
`/std:c11` flag explicitely. This is not done if an older Visual Studio
solution is upgraded to a newer Platform Toolset (e.g. going from Visual
Studio 2017 with v141 to Visual Studio 2019 with v142). Removing the
check for the version and always using the alternative stdatomic and
threads headers allows us to provide Visual Studio solutions for our
examples that are based on Visual Studio 2017 with toolset v141 and the
user can optionally update it to toolset v142 without having problems
with these includes.