`GVSPAdjustPacketSize` is part of `Stream` module
The feature was never executed in the old implementation, because the
wrong handle was used. This never resulted in an error, because the
comparison in the `if` statement was simply evaluated to `false` and the
execution continued.

TODO: this does not check return code of `VmbCameraInfoQuery`. This
should probably be done if we want this to be a good example for our
customers.